requirements.txt
py-cord==2.6.0
pymongo==4.8.0
python-dotenv
pytz==2024.1